Added startcpp script, for generating CMake project, added  in files that I forgot to do that previously.

+++ Scripts/rebuild.sh
@@ -1,17 +1,18 @@
 #!/bin/sh
 
-sudo nvim /etc/nixos/
+set -e
 
-if sudo nixos-rebuild switch --flake /etc/nixos#hostname; then
+if [[ $1 != "-v" ]]; then
+    sudo nvim /etc/nixos/
     cd ~/.config/nixos-config
+    if sudo nixos-rebuild switch --flake /etc/nixos#hostname; then
+        sudo rm -rf ./nixos
+        sudo cp -r /etc/nixos ./
 
-    sudo rm -rf ./nixos
-    sudo cp -r /etc/nixos ./
-
-    git add -A
-    git commit -m "Hello"
-    git push origin main
+        git add -A
+        git commit -m "Hello"
+        git push origin main
+    fi
+else
+    nvim /etc/nixos/
 fi
-
-
-
